requirements.txt
pandas
pyarrow
requests
beautifulsoup4
extract-favicon
torch
torchvision
scikit-learn
Pillow
cairosvg
urllib3

